requirements.txt
absl-py==1.4.0
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cycler==0.11.0
fastapi==0.103.1
h11==0.14.0
httptools
idna==3.4
kiwisolver==1.4.5
mediapipe==0.10.14
numpy==1.26.0
opencv-contrib-python==4.8.1.78
Pillow==10.0.1
protobuf==4.25.3   # Updated to be compatible with mediapipe 0.10.14
pydantic==2.3.0
pyparsing==3.0.9   # Compatible with matplotlib 3.7.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
starlette==0.27.0  # Compatible with fastapi 0.103.1
typing-extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.1
watchgod==0.8.2
websockets==11.0.3
